:recycle: Move bulk replace dataset entries to its own file.
