+++ api/src/models/dataset-integration.ts
@@ -15,7 +15,7 @@ import {
 import sequelize from "@/db/db-client"
 
 import Dataset from "@/models/dataset"
-import { activate, applyJMESPathTransform } from "@/models/dataset-integrations"
+import { activate, applyJMESPathTransform, bulkReplaceDatasetEntries } from "@/models/dataset-integrations"
 
 // Keep in sync with web/src/api/dataset-fields-api.ts
 export enum DatasetIntegrationStatusTypes {
@@ -88,6 +88,12 @@ export class DatasetIntegration extends Model<
     this: DatasetIntegration
   ): Promise<NonAttribute<DatasetIntegration>> {
     return applyJMESPathTransform(this)
+  }
+
+  async bulkReplaceDatasetEntries(
+    this: DatasetIntegration
+  ): Promise<void> {
+    return bulkReplaceDatasetEntries(this)
   }
 }
