+++ api/src/models/dataset-integrations/index.ts
@@ -1,2 +1,3 @@
 export { activate } from "./activate"
 export { applyJMESPathTransform } from "./apply-jmes-path-transform"
+export { bulkReplaceDatasetEntries } from "./bulk-replace-dataset-entries"
