+++ api/src/services/datasets/refresh/create-service.ts
@@ -1,9 +1,4 @@
-import { CreationAttributes } from "sequelize"
-import { isArray, isNil, isString } from "lodash"
-import jmespath from "jmespath"
-
-import db, { DatasetEntry, DatasetIntegration, User } from "@/models"
-import { DatasetEntryJsonDataType } from "@/models/dataset-entry"
+import db, { DatasetIntegration, User } from "@/models"
 
 import BaseService from "@/services/base-service"
 
@@ -23,67 +18,12 @@ export class CreateService extends BaseService {
 
       await this.datasetIntegration.applyJMESPathTransform()
       // TODO: create fields if none exist during dataset import
-      await this.bulkReplaceDatasetEntries(this.datasetIntegration)
+      await this.datasetIntegration.bulkReplaceDatasetEntries()
+
+      // TODO: log user action
+
       return this.datasetIntegration.save()
     })
-  }
-
-  private async parseJsonData(datasetIntegration: DatasetIntegration) {
-    const { rawJsonData, jmesPathTransform } = datasetIntegration
-
-    if (isNil(rawJsonData)) {
-      throw new Error("An integration must have data to be parsed.")
-    }
-
-    if (isNil(jmesPathTransform) && isArray(rawJsonData)) {
-      return datasetIntegration.set({
-        parsedJsonData: rawJsonData,
-      })
-    }
-
-    if (isNil(jmesPathTransform)) {
-      throw new Error("An integration must parse to an array to be valid")
-    }
-
-    const parsedJsonData = jmespath.search(rawJsonData, jmesPathTransform)
-    return datasetIntegration.set({
-      parsedJsonData,
-    })
-  }
-
-  private async bulkReplaceDatasetEntries(datasetIntegration: DatasetIntegration) {
-    const { datasetId, parsedJsonData } = datasetIntegration
-
-    await DatasetEntry.destroy({
-      where: {
-        datasetId,
-      },
-    })
-
-    if (isNil(parsedJsonData)) {
-      throw new Error("An integration must have data to build data entries.")
-    }
-
-    const datasetEntriesAttributes: CreationAttributes<DatasetEntry>[] = parsedJsonData.map(
-      (rawJsonData) => {
-        let jsonData: DatasetEntryJsonDataType
-        if (isString(rawJsonData)) {
-          jsonData = {
-            value: rawJsonData,
-          } as DatasetEntryJsonDataType
-        } else {
-          jsonData = rawJsonData as DatasetEntryJsonDataType
-        }
-
-        return {
-          datasetId,
-          rawJsonData,
-          jsonData,
-        }
-      }
-    )
-    await DatasetEntry.bulkCreate(datasetEntriesAttributes)
-    return
   }
 }
